backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
Pillow==10.0.0
stripe==5.0.0
django-filter==23.2.0
psycopg2-binary==2.9.6
django-environ==0.8.1
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.2.2